backend/pyproject.toml
[project]
name = "jit-backend"
version = "0.1.0"
description = "JIT Admin API"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn==0.35.0",
  "PyJWT==2.10.1",
  "kubernetes==30.1.0",
  "pydantic==2.11.7",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
